refactor(disaggregation): restructure into base/common/jax_transfer tiers
Reorganize the PD disaggregation package into a clear three-tier
dependency structure (base/ → common/ → jax_transfer/) following
sglang's established pattern. Consolidate 10 test files into 5 and
fix multi-host test scripts broken by the multi-entry payload API.
