+++ python/sgl_jax/srt/disaggregation/base/__init__.py
@@ -6,6 +6,7 @@ from sgl_jax.srt.disaggregation.base.kv_manager import (
     KVReceiver,
     KVSender,
     StateHolder,
+    TransferBackend,
     is_legal_transition,
 )
 
@@ -17,5 +18,6 @@ __all__ = [
     "LEGAL_TRANSITIONS",
     "TERMINAL_STATES",
     "StateHolder",
+    "TransferBackend",
     "is_legal_transition",
 ]
