+++ python/sgl_jax/srt/disaggregation/jax_transfer/__init__.py
@@ -1,13 +1,15 @@
+from sgl_jax.srt.disaggregation.common.zmq_notifier import ZmqPullNotifier
 from sgl_jax.srt.disaggregation.jax_transfer.conn import (
+    JaxTransferBackend,
     JaxTransferKVManager,
     JaxTransferKVReceiver,
     JaxTransferKVSender,
     PMetadata,
     TransferStatus,
 )
-from sgl_jax.srt.disaggregation.jax_transfer.zmq_notifier import ZmqPullNotifier
 
 __all__ = [
+    "JaxTransferBackend",
     "JaxTransferKVManager",
     "JaxTransferKVReceiver",
     "JaxTransferKVSender",
